+++ CHANGELOG.md
@@ -6,6 +6,10 @@ Features
 
 - Add average locktime metric to STAKE
 
+Bug fixes
+
+- Fixed an issue where the stake forms were not properly reset when changing account
+
 ## v1.0.0
 
 - Initial app release